[Refactor] Opt Column 3
