+++ be/src/base/simd/gather.h
@@ -20,6 +20,7 @@
 #include <immintrin.h>
 #endif
 
+#include <cstddef>
 #include <cstdint>
 
 namespace starrocks {
