+++ be/src/column/adaptive_nullable_column.cpp
@@ -16,9 +16,9 @@
 
 #include "base/simd/simd.h"
 #include "column/column_helper.h"
+#include "column/mysql_row_buffer.h"
 #include "gutil/casts.h"
 #include "gutil/strings/fastmem.h"
-#include "util/mysql_row_buffer.h"
 
 namespace starrocks {
