+++ be/src/column/decimalv3_column.cpp
@@ -16,6 +16,7 @@
 
 #include "base/types/int256.h"
 #include "column/fixed_length_column.h"
+#include "column/mysql_row_buffer.h"
 
 namespace starrocks {
