+++ be/src/column/german_string.cpp
@@ -15,10 +15,9 @@
 #include "column/german_string.h"
 
 #include "base/container/raw_container.h"
+#include "base/hash/hash_util.hpp"
 #include "base/string/slice.h"
 #include "gutil/strings/fastmem.h"
-#include "util/hash_util.hpp"
-#include "util/misc.h"
 
 namespace starrocks {
