+++ be/src/column/map_column.cpp
@@ -20,6 +20,7 @@
 #include "column/column_helper.h"
 #include "column/column_view/column_view.h"
 #include "column/fixed_length_column.h"
+#include "column/mysql_row_buffer.h"
 #include "column/nullable_column.h"
 #include "column/vectorized_fwd.h"
 #include "exec/sorting/sorting.h"
@@ -27,7 +28,6 @@
 #include "gutil/casts.h"
 #include "gutil/strings/fastmem.h"
 #include "types/datum.h"
-#include "util/mysql_row_buffer.h"
 
 namespace starrocks {
 void MapColumn::check_or_die() const {
