+++ be/src/column/nullable_column.cpp
@@ -17,10 +17,10 @@
 #include "base/simd/simd.h"
 #include "column/column_helper.h"
 #include "column/column_view/column_view.h"
+#include "column/mysql_row_buffer.h"
 #include "column/vectorized_fwd.h"
 #include "gutil/casts.h"
 #include "gutil/strings/fastmem.h"
-#include "util/mysql_row_buffer.h"
 
 namespace starrocks {
