+++ be/test/column/CMakeLists.txt
@@ -15,6 +15,8 @@
 set(COLUMN_TEST_FILES
     ../base/cxa_throw_wrap.cpp
     column_base_test.cpp
+    mysql_row_buffer_test.cpp
+    german_string_test.cpp
     column_visitor_test.cpp
     column_hash_stub.cpp
 )
