+++ be/test/column/mysql_row_buffer_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "util/mysql_row_buffer.h"
+#include "column/mysql_row_buffer.h"
 
 #include <optional>
